Cleanly align cat photos.
This commit employs some flex styling on the photos page to make sure
all of the cat photos are vertically and horizontally spaced in a clean
way.

We also add a new entry to the tmux launch and the npm scripts, so that
we can have the build pipeline set up to support a dynamically generated
stylesheet.