Fetch remote chat content from API
Just like with the remote News API in the previous commit, the chats
interface updates to use a remote API in this one.  This commit also
implements a naive caching system which makes sure we don't try to
double-fetch data which we've already asked for once.  Right now this
cache is memory-only (vs. like, local storage), so a page refresh will
destroy it.